requirements.txt
httpx==0.27.0
httpx-sse==0.4.0
cerebras_cloud_sdk==1.14.0
pydantic==2.9.2
instructor==1.7.0
deepgram-sdk==3.7.4
ffmpeg==1.4
groq==0.13.0
numpy==1.26.4
openai==1.55.3
pydub==0.25.1
python-dotenv==1.0.0
requests==2.32.3
sounddevice==0.5.1
soundfile==0.12.1
websocket-client==1.8.0
websockets==13.1
PyAudio==0.2.14
pydantic_core==2.23.4
pydantic-settings==2.6.1
absl-py==2.1.0
aenum==3.1.15
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncer==0.0.7
attrs==23.2.0
python-http-client==3.3.7
